# [MB-6628] Use EagerPreload to improve performance in some key spots