@reggieriser 2021-02-11
I've created a test case and filed an issue with Pop about the problems using `EagerPreload` with long association paths: https://github.com/gobuffalo/pop/issues/626

I've also added notes about the issue to our Wiki's [`EagerPreload` page](https://github.com/transcom/mymove/wiki/Using-EagerPreload-in-Pop#associations-with-3-path-elements-where-the-first-2-path-elements-match).